requirements.txt
# MERGES FASTAPI WITH MONGO DB

fastapi 
uvicorn
motor
pydantic
APScheduler
feedparser
python-dotenv
pymongo==4.3.3
pymongo[zstd]
pymongo[srv]

